modularizing website's mainsections
