+++ template-parts/footer/cta-footer.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Template part to display the CTA in all pages
+ * 
+ * @package Compara_inmuebles
+ * @since 1.0.0
+ */
+?>
+
+<!-- CALL TO ACTION START (call-to-action-6) -->
+<div class="ltn__call-to-action-area call-to-action-6 before-bg-bottom">
+    <div class="container">
+        <div class="row">
+            <div class="col-lg-12">
+                <div class="call-to-action-inner call-to-action-inner-6 ltn__secondary-bg text-center---">
+                    <div class="coll-to-info text-color-white">
+                        <h1>Looking for a dream home?</h1>
+                        <p>We can help you realize your dream of a new home</p>
+                    </div>
+                    <div class="btn-wrapper">
+                        <a class="btn btn-effect-3 btn-white" href="contact.html">Explore Properties <i class="icon-next"></i></a>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<!-- CALL TO ACTION END -->
